Added DomUtilTest. Clean up imports.
Added a test that disproves #54.

Closes #54
