+++ core/src/main/java/com/crawljax/core/CrawlerContext.java
@@ -88,7 +88,8 @@ public class CrawlerContext {
 	}
 
 	/**
-	 * @param The current {@link CrawlPath} in this context.
+	 * @param The
+	 *            current {@link CrawlPath} in this context.
 	 */
 	protected void setCrawlPath(CrawlPath path) {
 		crawlpath = path;
