+++ core/src/main/java/com/crawljax/core/plugin/PreCrawlingPlugin.java
@@ -1,5 +1,6 @@
 package com.crawljax.core.plugin;
 
+import java.net.URL;
 
 import com.crawljax.core.configuration.CrawljaxConfiguration;
